django-cloudrun/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Django==3.1.1
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.3.0
django-grappelli==2.14.2
django-jet==1.0.8
django-oauth-toolkit==1.3.2
django-storages==1.9.1
djangorestframework==3.11.1
drfdocs==0.0.11
google-api-core==1.22.2
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
